templates/generator/application/api-gateway/requirements.txt
pyzmq

opentelemetry-distro
opentelemetry-exporter-otlp-proto-http
